except 3 *

+++ controllers/blogs.js
@@ -1,14 +1,15 @@
 const blogsRouter = require('express').Router()
-const { response } = require('../app')
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async(request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs)
 })
 
 blogsRouter.get('/:id', async(request, response) => {
-    const blog = await Blog.findById(request.params.id)
+    const blog = await Blog.findById(request.params.id).populate('user')
     if (blog) {
         response.json(blog)
     } else {
@@ -18,19 +19,41 @@ blogsRouter.get('/:id', async(request, response) => {
 
 blogsRouter.post('/', async(request, response) => {
     const body = request.body
+    const user = request.user
+
+    if (!user) {
+        return response.status(401).json({ error: 'token is missing or invalid' })
+    }
 
     const blog = new Blog({
         title: body.title,
         author: body.author,
         url: body.url,
-        likes: body.likes || 0
+        likes: body.likes || 0,
+        user: user._id,
     })
 
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+
     response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async(request, response) => {
+    const user = request.user
+    const blog = await Blog.findById(request.params.id)
+
+    if (!user) {
+        return response.status(401).json({ error: 'token is missing or invalid' })
+    }
+
+    if ( blog.user.toString() !== user.id.toString() ) {
+        return response.status(400).json({
+            error: 'token does not allow deleting this blog',
+        })
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
 })
